+++ command/agent/agent_test.go
@@ -1007,3 +1007,97 @@ func TestAgent_NodeMaintenanceMode(t *testing.T) {
 		t.Fatalf("bad: %#v", check)
 	}
 }
+
+func TestAgent_checkStateSnapshot(t *testing.T) {
+	config := nextConfig()
+	dir, agent := makeAgent(t, config)
+	defer os.RemoveAll(dir)
+	defer agent.Shutdown()
+
+	// First register a service
+	svc := &structs.NodeService{
+		ID:      "redis",
+		Service: "redis",
+		Tags:    []string{"foo"},
+		Port:    8000,
+	}
+	if err := agent.AddService(svc, nil, false); err != nil {
+		t.Fatalf("err: %v", err)
+	}
+
+	// Register a check
+	check1 := &structs.HealthCheck{
+		Node:        config.NodeName,
+		CheckID:     "service:redis",
+		Name:        "redischeck",
+		Status:      structs.HealthPassing,
+		ServiceID:   "redis",
+		ServiceName: "redis",
+	}
+	if err := agent.AddCheck(check1, nil, true); err != nil {
+		t.Fatalf("err: %s", err)
+	}
+
+	// Snapshot the state
+	snap := agent.snapshotCheckState()
+
+	// Unload all of the checks
+	if err := agent.unloadChecks(); err != nil {
+		t.Fatalf("err: %s", err)
+	}
+
+	// Reload the checks
+	if err := agent.loadChecks(config); err != nil {
+		t.Fatalf("err: %s", err)
+	}
+
+	// Restore the state
+	agent.restoreCheckState(snap)
+
+	// Search for the check
+	out, ok := agent.state.Checks()[check1.CheckID]
+	if !ok {
+		t.Fatalf("check should have been registered")
+	}
+
+	// Make sure state was restored
+	if out.Status != structs.HealthPassing {
+		t.Fatalf("should have restored check state")
+	}
+}
+
+func TestAgent_loadChecks_checkFails(t *testing.T) {
+	config := nextConfig()
+	dir, agent := makeAgent(t, config)
+	defer os.RemoveAll(dir)
+	defer agent.Shutdown()
+
+	// Persist a health check with an invalid service ID
+	check := &structs.HealthCheck{
+		Node:      config.NodeName,
+		CheckID:   "service:redis",
+		Name:      "redischeck",
+		Status:    structs.HealthPassing,
+		ServiceID: "nope",
+	}
+	if err := agent.persistCheck(check, nil); err != nil {
+		t.Fatalf("err: %s", err)
+	}
+
+	// Check to make sure the check was persisted
+	checkHash := stringHash(check.CheckID)
+	checkPath := filepath.Join(config.DataDir, checksDir, checkHash)
+	if _, err := os.Stat(checkPath); err != nil {
+		t.Fatalf("err: %s", err)
+	}
+
+	// Try loading the checks from the persisted files
+	if err := agent.loadChecks(config); err != nil {
+		t.Fatalf("err: %s", err)
+	}
+
+	// Ensure the erroneous check was purged
+	if _, err := os.Stat(checkPath); err == nil {
+		t.Fatalf("should have purged check")
+	}
+}