+++ terraform/openstack/scripts/service.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+set -e
+
+echo "Starting Consul..."
+sudo start consul